rename project to project-card
